Window_ShopParty: fix msvc size_t -> int warnings

+++ src/window_shopparty.cpp
@@ -96,7 +96,7 @@ void Window_ShopParty::Refresh() {
 		return;
 
 	const std::vector<Game_Actor*>& actors = Main_Data::game_party->GetActors();
-	for (size_t i = 0; i < actors.size() && i < 4; i++) {
+	for (int i = 0; i < static_cast<int>(actors.size()) && i < 4; i++) {
 		Game_Actor *actor = actors[i];
 		int phase = (cycle / anim_rate) % 4;
 		if (phase == 3) {
